% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pedigreeSim.R
\name{compare_phases}
\alias{compare_phases}
\title{Codifying phases according with GUSMaap
for comparisons}
\usage{
compare_phases(founderfile, ref_alt_alleles, filename = "simulated_phases.txt")
}
\arguments{
\item{founderfile}{input file for PedigreeSim software}

\item{ref_alt_alleles}{data.frame with chromosome, position, reference allele,
alternative allele, position in centimorgan information}
}
\description{
Codifying phases according with GUSMaap
for comparisons
}
