% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pedigreeSim.R
\name{create_chromfile}
\alias{create_chromfile}
\title{Creates chrom file PedigreeSim input
Only for diploid species}
\usage{
create_chromfile(mapfile, filename = "chromosome.txt")
}
\arguments{
\item{mapfile}{output object from create_mapfile}

\item{filename}{output file name}
}
\description{
Creates chrom file PedigreeSim input
Only for diploid species
}
