% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pedigreeSim.R
\name{create_haplo}
\alias{create_haplo}
\title{Creates PedigreeSim input files from empirical non-phased VCF file
The same alleles and proportion of marker types are kept from the VCF and the haplotypes are simulated
This function needs adaptation to deal with other populations and ploidy
It is important that MNP markers are in splitted format}
\usage{
create_haplo(vcfR.obj, ref.map, seed, P1, P2, filename = "founders.txt")
}
\arguments{
\item{vcfR.obj}{object of class vcfR}

\item{ref.map}{data.frame with information from reference linkage maps within columns bp and cm}

\item{seed}{integer defining the simulation seed}

\item{P1}{parent1 ID in VCF file}

\item{P2}{parent2 ID in VCF file}

\item{filename}{output file name}
}
\description{
Creates PedigreeSim input files from empirical non-phased VCF file
The same alleles and proportion of marker types are kept from the VCF and the haplotypes are simulated
This function needs adaptation to deal with other populations and ploidy
It is important that MNP markers are in splitted format
}
