% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pedigreeSim.R
\name{create_mapfile}
\alias{create_mapfile}
\title{Creates map file for PedigreeSim using a reference linkage map or a cM by Mb rate.}
\usage{
create_mapfile(
  vcfR.obj,
  ref.map = NULL,
  cMbyMb = NULL,
  filename = "mapfile.txt",
  rm_outlier = T,
  thr = 15
)
}
\arguments{
\item{vcfR.obj}{object of class vcfR}

\item{ref.map}{data.frame with two columns nameed cM and bp with numerical values}

\item{cMbyMb}{numeric value for the cM by Mb rate}
}
\description{
Creates map file for PedigreeSim using a reference linkage map or a cM by Mb rate.
}
