% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pedigreeSim.R
\name{create_parfile}
\alias{create_parfile}
\title{Creates parfile PedigreeSim input
By now only available for F1 diploid population}
\usage{
create_parfile(
  seed,
  popsize,
  filename = "parameters.txt",
  name.mapfile = "mapfile.txt",
  name.founderfile = "founderfile.txt",
  name.chromfile = "chromosome.txt",
  name.out = "sim"
)
}
\arguments{
\item{seed}{integer defining the simulation seed}

\item{popsize}{integer defining the population size}

\item{filename}{a character with the parfile name}

\item{name.mapfile}{a character with the mapfile name}

\item{name.founderfile}{a character with the founderfile name}

\item{name.chromfile}{a character with the chromfile name}

\item{name.out}{a character with the PedigreeSim outputs names}
}
\description{
Creates parfile PedigreeSim input
By now only available for F1 diploid population
}
