% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermassa_genotype.R
\name{depth_prepare}
\alias{depth_prepare}
\title{Creates list with first level refering to markers and second level with
marker name, offspring names and alternative and reference allele counts.
This function was created to parallelize the SuperMassa process. Its output
is the input of supermassa_error}
\usage{
depth_prepare(extracted_depth)
}
\description{
Creates list with first level refering to markers and second level with
marker name, offspring names and alternative and reference allele counts.
This function was created to parallelize the SuperMassa process. Its output
is the input of supermassa_error
}
