% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HPC_info.R
\name{efficiency_graphics}
\alias{efficiency_graphics}
\title{Get RAM and CPU usage from SLURM managment system}
\usage{
efficiency_graphics(eff_table, out_name)
}
\description{
Get RAM and CPU usage from SLURM managment system
}
\examples{
\dontrun{
  # grep -rw "Submitted batch job" * > jobs
  df <- read.table("/scratch/user/chtaniguti/cromwell-executions/SNPCalling/7cefff16-8d18-46d7-adeb-b22e4f35b663/jobs")
  df <- read.table("jobs_acca")
  tasks <- sapply(strsplit(df$V1, "/"), function(x) paste0(x[grep("call-", x)], collapse = "/"))
  tasks <- gsub("call-", "", tasks)
  jobs_acca <- data.frame(tasks, df$V4)
  write.table(df$V4, quote = F, file = "jobs_acca_id", col.names = F, row.names = F)

  df <- read.table("jobs_euc") 
  tasks <- sapply(strsplit(df$V1, "/"), function(x) paste0(x[grep("call-", x)], collapse = "/"))

  tasks <- gsub("call-", "", tasks)
  jobs_euc <- data.frame(tasks, df$V4)
  write.table(df$V4, quote = F, file = "jobs_euc_id", col.names = F, row.names = F)

  # for i in $(cat jobs_euc_id);do
  #   seff $i >> infos.temp.euc
  # done

  df.acca <- efficiency_table(infos_seff = "infos.temp.acca", jobs_acca)
  efficiency_graphics(df.acca, out_name = "acca_efficiency_wf.png")

  df.euc <- efficiency_table(infos_seff = "infos.temp.euc", jobs_euc)
  efficiency_graphics(df.euc, out_name = "euc_efficiency_wf.png")
}

}
