% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_multiallelic_vcf.R
\name{get_alternatives}
\alias{get_alternatives}
\title{Replace the reference and alternative alleles in fix to 
keep only the ones found in the parents}
\usage{
get_alternatives(fix, filt.gt, P1, P2)
}
\arguments{
\item{fix}{fix part of vcfR object}

\item{P1}{character with one of the parents ID}

\item{P2}{character with other parent ID}

\item{gt}{genotypes matrix (markers x individuals) with multiallelic markers}
}
\description{
Replace the reference and alternative alleles in fix to 
keep only the ones found in the parents
}
