% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onemap_write_vcfR.R
\name{onemap_write_vcfR}
\alias{onemap_write_vcfR}
\title{Write VCF files from onemap objects}
\usage{
onemap_write_vcfR(
  onemap.object,
  out_vcf,
  input_info_rds,
  parent1.geno = NULL,
  parent1.id = NULL,
  parent2.geno = NULL,
  parent2.id = NULL,
  probs,
  ad_matrix = NULL
)
}
\arguments{
\item{onemap.object}{object of onemap class}

\item{out_vcf}{path and name of the vcf file to be outputed from the generated onemap object. 
It is important to notice that only onemap informative markers are kept.}

\item{input_info_rds}{output file from onemap_read_vcfR to store alleles information}

\item{parent1.geno}{parent 1 genotypes (vcf notation: 0/0, 0/1, 1/1, ./.)}

\item{parent1.id}{parent 1 ID}

\item{parent2.geno}{parent 2 genotypes (vcf notation: 0/0, 0/1, 1/1, ./.)}

\item{parent2.id}{parent 2 ID}

\item{probs}{genotyope probabilities}
}
\value{
vcf file
}
\description{
Only available for biallelic markers. OneMap don't keep the reference and alternative 
allele information. Do not consider the outputted information of the alleles for further 
investigations. We can't properly relate homozygous progeny from both parents heterozygous 
with the reference (0/0) or alternative (1/1) alleles.
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
