% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyRAD_genotype.R
\name{polyRAD_genotype_vcf}
\alias{polyRAD_genotype_vcf}
\title{Uses polyRAD RADdata2VCF and  Export_MAPpoly to generate the VCF file
The probabilities exported by Export_MAPpoly are includede in the 
RADdata2VCF vcf file}
\usage{
polyRAD_genotype_vcf(vcf, parent1, parent2, outfile = "out.vcf.gz")
}
\arguments{
\item{vcf}{vcf file}

\item{parent1}{parent 1 identification in vcfR object}

\item{parent2}{parent 2 identification in vcfR objetc}

\item{outfile}{output vcf file name}
}
\description{
Uses polyRAD RADdata2VCF and  Export_MAPpoly to generate the VCF file
The probabilities exported by Export_MAPpoly are includede in the 
RADdata2VCF vcf file
}
