% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_non_informative.R
\name{remove_non_informative}
\alias{remove_non_informative}
\title{Filter non-informative markers. Accepts only biallelic markers.}
\usage{
remove_non_informative(
  vcf,
  P1,
  P2,
  out.vcf = "filtered.vcf.gz",
  replaceAD = TRUE
)
}
\arguments{
\item{vcf}{path to the vcf file}

\item{P1}{parent 1 ID}

\item{P2}{parent 2 ID}

\item{out.vcf}{output VCF file name}

\item{replaceAD}{logical, if true it replaces AD by missing data when GT is missing}
}
\description{
Filter non-informative markers. Accepts only biallelic markers.
}
