% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pedigreeSim.R
\name{remove_outlier}
\alias{remove_outlier}
\title{Remove inverted SNPs
Sometimes the input linkage map has inversions compared with the physical position,
Here these markers are eliminated}
\usage{
remove_outlier(ref.map, thr = 15)
}
\arguments{
\item{ref.map}{data.frame with columns bp and cM}

\item{thr}{measure in centimorgan defining the maximum difference accepted for inversions}
}
\description{
Remove inverted SNPs
Sometimes the input linkage map has inversions compared with the physical position,
Here these markers are eliminated
}
