% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pedsim.R
\name{run_pedsim}
\alias{run_pedsim}
\title{Wrapper function for PedigreeSim software}
\usage{
run_pedsim(
  chromosome = c("Chr1", "Chr2", "Chr3"),
  n.marker = c(18, 19, 20),
  tot.size.cm = c(18, 19, 20),
  ref.vcf = NULL,
  ref.map = NULL,
  centromere = c(9, 9, 10),
  n.ind = 150,
  mk.types = c("A1", "D1.10", "D2.15"),
  n.types = c(7, 25, 25),
  pop = c("F1", "F2", "BC"),
  seed = 8080,
  name.mapfile = "mapfile.map",
  name.founderfile = "founderfile.gen",
  name.chromfile = "sim.chrom",
  name.parfile = "sim.par",
  name.out = "sim_out",
  rm.tempfiles = F,
  parent1 = NULL,
  parent2 = NULL
)
}
\arguments{
\item{chromosome}{a vector of characters defining the chromosomes names}

\item{n.marker}{a vector of integers defining the number of markers within chromosome}

\item{tot.size.cm}{a vector of double numbers defining the total size in centimorgan of each chromosome}

\item{ref.vcf}{object of class vcfR}

\item{ref.map}{data.frame with position in base pair and in centimorgan (reference linkage map)}

\item{centromere}{a vector of double numbers defining the centromere position}

\item{n.ind}{a integer defining the number of individuals to be simulated in the progeny}

\item{mk.types}{a vector of characters defining the marker types according with Wu2002a
(the types will be distribuited randomly by the chromosomes)}

\item{n.types}{a vector of integers defining the number of markers of each type}

\item{pop}{a character defining the population type. Supported: "F1", "F2", "BC"}

\item{seed}{integer defining the simulation seed}

\item{name.mapfile}{a character with the mapfile name}

\item{name.founderfile}{a character with the founderfile name}

\item{name.chromfile}{a character with the chromfile name}

\item{name.parfile}{a character with the parfile name}

\item{name.out}{a character with the PedigreeSim outputs names}

\item{rm.tempfiles}{logical TRUE/FALSE, if TRUE remove the temporary files generated}

\item{parent1}{parent1 ID defined in reference VCF}

\item{parent2}{parent2 ID defined in reference VCF}

\item{path.pedsim}{path to the .jar PedigreeSim file. If you still don't have it,
download in https://www.wur.nl/en/show/Software-PedigreeSim.htm}
}
\value{
PedigreeSim outputs including the population genotypes codified as the observed bands genotypes
according to Wu et. al 2002 table I if a VCF file is not included. If a reference VCF is included, the genotypes
in output are accordind to the reference and alternative allele defined in VCF.
}
\description{
With the function arguments users can define the basic characteristics
of the maps to be simulated or users can provide a VCF and a reference linkage map files
to define the number and type of markers, and genetic distances.
The PedigreeSim software offers other options that are not available here.
}
\examples{
\dontrun{
# For outcross
run_pedsim(chromosome = c("Chr1", "Chr2"), n.marker = c(40,32), tot.size.cm = c(100,150), centromere = c(50, 75),
           n.ind = 200, mk.types = c("A1", "A2", "A3", "A4", "B1.5", "B2.6", "B3.7",
                                     "C.8", "D1.9", "D1.10", "D1.11", "D1.12", "D1.13",
                                     "D2.14", "D2.15", "D2.16", "D2.17", "D2.18"),
                                     n.types = rep(4,18), pop = "F1", path.pedsim = "~/",
           name.mapfile = "mapfile.map", name.founderfile="founderfile.gen", name.chromfile="sim.chrom", name.parfile="sim.par",
           name.out="sim_out")
}

}
\references{
Wu, R., Ma, C.-X., Painter, I. and Zeng, Z.-B. (2002)
Simultaneous maximum likelihood estimation of linkage and linkage phases in
outcrossing species. \emph{Theoretical Population Biology} 61: 349-363.
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
