% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregation_test_vcf.R
\name{segregation_test_vcf}
\alias{segregation_test_vcf}
\title{Filter markers with segregation distortion. Accepts only biallelic markers.}
\usage{
segregation_test_vcf(vcf, threshold, P1, P2, out.vcf = "filtered.vcf.gz")
}
\arguments{
\item{vcf}{path to the vcf file}

\item{threshold}{alpha value to be considered}

\item{P1}{parent 1 ID}

\item{P2}{parent 2 ID}

\item{out.vcf}{output VCF file name}
}
\description{
Filter markers with segregation distortion. Accepts only biallelic markers.
}
