% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermassa_genotype.R
\name{supermassa_genotype_vcf}
\alias{supermassa_genotype_vcf}
\title{Runs SuperMassa for multiple markers using doParallel packages
Updates OneMap object in genotypes, error probabilitie and marker type
Also removes non-informative markers according to SuperMassa genotyping.}
\usage{
supermassa_genotype_vcf(
  vcf = NULL,
  vcf.par = c("AD", "DPR"),
  out_vcf = NULL,
  parent1 = "P1",
  parent2 = "P2",
  crosstype = NULL,
  cores = 2,
  ploidy = NULL
)
}
\arguments{
\item{vcf}{path and name of the vcf file}

\item{vcf.par}{Field of VCF that informs the depth of alleles}

\item{out_vcf}{path and name of the vcf file to be outputed from the generated onemap object. 
It is important to notice that only onemap informative markers are kept.}

\item{parent1}{parent 1 identification in vcfR object}

\item{parent2}{parent 2 identification in vcfR objetc}

\item{crosstype}{string defining the cross type, by now it supports only 
outcross and f2 intercross}

\item{cores}{number of threads}

\item{f1}{f1 individual identification if f2 cross type}
}
\description{
Runs SuperMassa for multiple markers using doParallel packages
Updates OneMap object in genotypes, error probabilitie and marker type
Also removes non-informative markers according to SuperMassa genotyping.
}
