% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermassa_genotype.R
\name{supermassa_parallel_poly}
\alias{supermassa_parallel_poly}
\title{Use output list from depth_prepare and run Supermassa for each marker (first level of the list)
Extract from the SuperMassa output, the offspring genotype, the marker type, and the error
probabilities}
\usage{
supermassa_parallel_poly(supermassa_4parallel)
}
\description{
Use output list from depth_prepare and run Supermassa for each marker (first level of the list)
Extract from the SuperMassa output, the offspring genotype, the marker type, and the error
probabilities
}
