% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_genotype.R
\name{updog_genotype_vcf}
\alias{updog_genotype_vcf}
\title{Run updog f1 model using allele depth from a VCF file and export the result also as VCF}
\usage{
updog_genotype_vcf(
  vcf = NULL,
  vcf.par = c("AD", "DPR"),
  out_vcf = NULL,
  parent1 = "P1",
  parent2 = "P2",
  crosstype = NULL,
  cores = 2,
  ploidy = NULL
)
}
\arguments{
\item{vcf}{path and name of the vcf file}

\item{vcf.par}{Field of VCF that informs the depth of alleles}

\item{out_vcf}{path and name of the vcf file to be outputed from the generated onemap object. 
It is important to notice that only onemap informative markers are kept.}

\item{parent1}{parent 1 identification in vcfR object}

\item{parent2}{parent 2 identification in vcfR objetc}

\item{crosstype}{string defining the cross type, by now it supports only 
outcross and f2 intercross}

\item{cores}{number of threads}
}
\value{
a VCF file
}
\description{
Uses alelle counts to reestimate genotypes with updog approach and 
stores the genotypes probabilities or for further multipoint 
analysis. This function only works with biallelic markers.
}
\references{
Gerard, D., Ferrão L.F.V., Garcia, A.A.F., & Stephens, M. (2018). Harnessing 
Empirical Bayes and Mendelian Segregation for Genotyping Autopolyploids from 
Messy Sequencing Data. bioRxiv. doi: 10.1101/281550.
}
\seealso{
\url{https://github.com/dcgerard/updog}.
}
\author{
Cristiane Taniguti, \email{chtaniguti@tamu.edu}
}
