% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2onemap.R
\name{vcf2onemap}
\alias{vcf2onemap}
\title{Wrapper function for Scott Brainard java function converter.jar}
\usage{
vcf2onemap(
  vcf_file = NULL,
  data_type = c("f2 backcross", "f2 intercross", "ri self", "ri sib", "outcross"),
  parent1 = NULL,
  parent2 = NULL,
  outputfile = NULL,
  verify_uniform_offspring = TRUE,
  log_filtered_markers = FALSE,
  only_phased = FALSE,
  java.converter.path = NULL
)
}
\arguments{
\item{vcf_file}{absolute path of the VCF file (can be gzipped)}

\item{data_type}{cross type (f2 backcross, f2 intercross, ri self, ri sib, or outcross)}

\item{parent1}{ID of the parent 1, as encoded in the VCF file}

\item{parent2}{ID of the parent 2, as encoded in the VCF file}

\item{outputfile}{absolute path to use for the OneMap data file and .log files}

\item{verify_uniform_offspring}{check all aa x aa and aa x bb segregation types, and record violations of
expected progeny genotypes (i.e., "aa" and "ab", respectively), in the .log file}

\item{log_filtered_markers}{false by default. Even when false, some summary statistics about how many markers
were filtered and why. If set to true, separate .log files will be generated for markers that are filtered due to:}

\item{only_phased}{retain only markers that are already phased in the VCF file (true/false, default=false)}

\item{java.converter.path}{path to converter.jar script}
}
\description{
The java script convert the VCF file to onemap raw file
}
