% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermassa_genotype.R
\name{supermassa_genotype}
\alias{supermassa_genotype}
\title{Runs SuperMassa for multiple markers using doParallel packages
Updates OneMap object in genotypes, error probabilitie and marker type
Also removes non-informative markers according to SuperMassa genotyping.}
\usage{
supermassa_genotype(
  vcfR.object = NULL,
  onemap.object = NULL,
  vcf.par = c("AD", "DPR"),
  parent1 = "P1",
  parent2 = "P2",
  f1 = NULL,
  recovering = FALSE,
  mean_phred = 20,
  cores = 2,
  depths = NULL,
  global_error = NULL,
  use_genotypes_errors = TRUE,
  use_genotypes_probs = FALSE,
  rm_multiallelic = TRUE
)
}
\arguments{
\item{vcfR.object}{object output of the vcfR package}

\item{onemap.object}{object of class onemap}

\item{vcf.par}{Field of VCF that informs the depth of alleles}

\item{parent1}{parent 1 identification in vcfR object}

\item{parent2}{parent 2 identification in vcfR objetc}

\item{f1}{f1 individual identification if f2 cross type}

\item{recovering}{logical defining if markers should be recovered from VCF}

\item{mean_phred}{the mean phred score of the sequencer technology}

\item{cores}{number of threads}

\item{depths}{list containing a matrix for ref and other for alt allele counts, samples ID in colum and markers ID in rows}

\item{global_error}{number from 0 to 1 defining the global error to be considered together 
with the genotype errors or the genotype probabilities or NULL to not considered any global error}

\item{use_genotypes_errors}{if \code{TRUE} the error probability of each genotype will be considered in emission function of HMM}

\item{use_genotype_probs}{if \code{TRUE} the probability of each possible genotype will be considered in emission function of HMM}
}
\description{
Runs SuperMassa for multiple markers using doParallel packages
Updates OneMap object in genotypes, error probabilitie and marker type
Also removes non-informative markers according to SuperMassa genotyping.
}
